requirements.txt
pyqt5
tqdm
Pillow
hachoir